Update benchmark with top-sites update
